requirements.txt
numpy
PyYAML
wandb
Pillow
setuptools
thop
wandb